+++ stack/stack.py
@@ -1,131 +1,6 @@
-class Node:
-    def __init__(self, value=None, next_node=None):
-        # the value at this linked list node
-        self.value = value
-        # reference to the next node in the list
-        self.next_node = next_node
-
-    def get_value(self):
-        return self.value
-
-    def get_next(self):
-        return self.next_node
-
-    def set_next(self, new_next):
-        # set this node's next_node reference to the passed in node
-        self.next_node = new_next
-
-class LinkedList:
-    def __init__(self):
-        # reference to the head of the list
-        self.head = None
-        # reference to the tail of the list
-        self.tail = None
-
-    def add_head(self, value):
-        new_node = Node(value)
-        if self.head:
-            new_node.next_node = self.head # move the current head to the next node
-            self.head = new_node # make the head the new node
-        else:
-            self.head = new_node
-            self.tail = new_node
-
-    def add_to_tail(self, value):
-        # wrap the input value in a node
-        new_node = Node(value, None)
-        # check if there is no head (i.e., the list is empty)
-        if not self.head:
-            # if the list is initially empty, set both head and tail to the new node
-            self.head = new_node
-            self.tail = new_node
-        # we have a non-empty list, add the new node to the tail
-        else:
-            # set the current tail's next reference to our new node
-            self.tail.set_next(new_node)
-            # set the list's tail reference to the new node
-            self.tail = new_node
-
-    def remove_head(self):
-        # return None if there is no head (i.e. the list is empty)
-        if not self.head:
-            return None
-        # if head has no next, then we have a single element in our list
-        if not self.head.get_next():
-            # get a reference to the head
-            head = self.head
-            # delete the list's head reference
-            self.head = None
-            # also make sure the tail reference doesn't refer to anything
-            self.tail = None
-            # return the value
-            return head.get_value()
-        # otherwise we have more than one element in our list
-        value = self.head.get_value()
-        # set the head reference to the current head's next node in the list
-        self.head = self.head.get_next()
-        return value
-
-    def remove_tail(self):
-        if not self.head:
-            return None
-        
-        if self.head is self.tail:
-            value = self.head.get_value()
-            self.head = None
-            self.tail = None
-            return value
-        
-        current = self.head
-
-        while current.get_next() is not self.tail:
-            current = current.get_next()
-
-        value = self.tail.get_value()
-        self.tail = current
-        return value
-
-    def contains(self, value):
-        if not self.head:
-            return False
-
-        # Recursive solution
-        # def search(node):
-        #   if node.get_value() == value:
-        #     return True
-        #   if not node.get_next():
-        #     return False
-        #   return search(node.get_next())
-        # return search(self.head)
-    
-        # get a reference to the node we're currently at; update this as we traverse the list
-        current = self.head
-        # check to see if we're at a valid node 
-        while current:
-            # return True if the current value we're looking at matches our target value
-            if current.get_value() == value:
-                return True
-            # update our current node to the current node's next node
-            current = current.get_next()
-        # if we've gotten here, then the target node isn't in our list
-        return False
-
-    def get_max(self):
-        if not self.head:
-            return None
-        # reference to the largest value we've seen so far
-        max_value = self.head.get_value()
-        # reference to our current node as we traverse the list
-        current = self.head.get_next()
-        # check to see if we're still at a valid list node
-        while current:
-            # check to see if the current value is greater than the max_value
-            if current.get_value() > max_value:
-                # if so, update our max_value variable
-                max_value = current.get_value()
-            # update the current node to the next node in the list
-            current = current.get_next()
-        return max_value
+import sys
+sys.path.append('./singly_linked_list')
+from singly_linked_list import LinkedList
 
 """
 A stack is a data structure whose primary purpose is to store and
